Split Core run integrity quality boundaries
