+++ docs/CODE_QUALITY_FILE_GRADES.md
@@ -10,7 +10,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | `scripts` | 23 | 4200 | 99 | A+ | `smoke.sh` A+<br>`native-desktop-smoke.sh` A+<br>`packaged_window.py` A+ |
 | `source:QuillCodeAgent` | 70 | 8758 | 100 | A+ | `Agent.swift` A+<br>`TrustedRouterOAuth.swift` A+<br>`FlailDetector.swift` A+ |
 | `source:QuillCodeApp` | 474 | 54991 | 100 | A+ | `WorkspaceHTMLTranscriptRenderer.swift` A+<br>`QuillCodeReviewSurface.swift` A+<br>`QuillCodeSettingsSurface.swift` A+ |
-| `source:QuillCodeCore` | 45 | 5250 | 99 | A+ | `RunSpendFusePolicy.swift` A+<br>`RunIntegrityLexicons.swift` A+<br>`AutomationModels.swift` A+ |
+| `source:QuillCodeCore` | 51 | 5163 | 99 | A+ | `AutomationModels.swift` A+<br>`AppConfig.swift` A+<br>`RunSpendLedger.swift` A+ |
 | `source:QuillCodePersistence` | 8 | 819 | 100 | A+ | `PermissionRuleStore.swift` A+<br>`ConfigStore.swift` A+<br>`SecretStore.swift` A+ |
 | `source:QuillCodeSafety` | 13 | 2145 | 99 | A+ | `Safety.swift` A+<br>`PermissionRuleGatedSafetyReviewer.swift` A+<br>`StaticSafetyReadOnlyShellPolicy.swift` A+ |
 | `source:QuillCodeTools` | 134 | 14803 | 100 | A+ | `GitToolExecutor.swift` A+<br>`ShellToolExecutor.swift` A+<br>`GitPatchToolExecutor.swift` A+ |
@@ -32,9 +32,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 
 | Grade | Score | Lines | Module | File | Main Issues |
 | --- | ---: | ---: | --- | --- | --- |
-| A+ | 96 | 173 | `source:QuillCodeCore` | `Sources/QuillCodeCore/RunSpendFusePolicy.swift` | 1 lines >120 chars; 16 public declarations |
 | A+ | 97 | 635 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParitySwiftInteractionTargetAuditSupport.swift` | 9 lines >120 chars; 1 lines >160 chars |
-| A+ | 97 | 614 | `source:QuillCodeCore` | `Sources/QuillCodeCore/RunIntegrityLexicons.swift` | watch file size (614 lines) |
 | A+ | 97 | 529 | `test:QuillCodeAppTests` | `Tests/QuillCodeAppTests/WorkspaceReviewIntegrationTests.swift` | 16 lines >120 chars |
 | A+ | 97 | 516 | `scripts` | `scripts/smoke.sh` | watch file size (516 lines) |
 | A+ | 97 | 457 | `scripts` | `scripts/native-desktop-smoke.sh` | watch file size (457 lines) |
@@ -72,6 +70,8 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | A+ | 97 | 185 | `source:QuillCodeTools` | `Sources/QuillCodeTools/SkillResolver.swift` | 14 public declarations |
 | A+ | 97 | 182 | `source:QuillCodeApp` | `Sources/QuillCodeApp/QuillCodeSidebarFilterSurface.swift` | 30 public declarations |
 | A+ | 97 | 181 | `source:QuillCodeSafety` | `Sources/QuillCodeSafety/PermissionRuleGatedSafetyReviewer.swift` | 13 public declarations |
+| A+ | 97 | 179 | `source:QuillCodeApp` | `Sources/QuillCodeApp/QuillCodeThreadSidebarSurface.swift` | 20 public declarations |
+| A+ | 97 | 173 | `source:QuillCodeCore` | `Sources/QuillCodeCore/RunSpendLedger.swift` | 22 public declarations |
 
 ## Every File
 
@@ -725,13 +725,19 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | A+ | 100 | 117 | `source:QuillCodeCore` | `Sources/QuillCodeCore/ProjectInstructionDiagnosticResolutions.swift` | no automated issues |
 | A+ | 97 | 100 | `source:QuillCodeCore` | `Sources/QuillCodeCore/ProjectRef.swift` | 14 public declarations |
 | A+ | 100 | 21 | `source:QuillCodeCore` | `Sources/QuillCodeCore/QuillCodeNotificationPreferences.swift` | no automated issues |
-| A+ | 97 | 614 | `source:QuillCodeCore` | `Sources/QuillCodeCore/RunIntegrityLexicons.swift` | watch file size (614 lines) |
+| A+ | 100 | 41 | `source:QuillCodeCore` | `Sources/QuillCodeCore/RunIntegrityLexicons.swift` | no automated issues |
 | A+ | 100 | 66 | `source:QuillCodeCore` | `Sources/QuillCodeCore/RunIntegrityRecord.swift` | no automated issues |
 | A+ | 100 | 87 | `source:QuillCodeCore` | `Sources/QuillCodeCore/RunIntegrityReport.swift` | no automated issues |
 | A+ | 100 | 277 | `source:QuillCodeCore` | `Sources/QuillCodeCore/RunIntegrityScanner.swift` | no automated issues |
-| A+ | 96 | 173 | `source:QuillCodeCore` | `Sources/QuillCodeCore/RunSpendFusePolicy.swift` | 1 lines >120 chars; 16 public declarations |
+| A+ | 100 | 49 | `source:QuillCodeCore` | `Sources/QuillCodeCore/RunSpendFuseModels.swift` | no automated issues |
+| A+ | 99 | 125 | `source:QuillCodeCore` | `Sources/QuillCodeCore/RunSpendFusePolicy.swift` | 1 lines >120 chars |
 | A+ | 97 | 173 | `source:QuillCodeCore` | `Sources/QuillCodeCore/RunSpendLedger.swift` | 22 public declarations |
 | A+ | 100 | 83 | `source:QuillCodeCore` | `Sources/QuillCodeCore/SubagentModels.swift` | no automated issues |
+| A+ | 100 | 37 | `source:QuillCodeCore` | `Sources/QuillCodeCore/SuccessClaimLexicon.swift` | no automated issues |
+| A+ | 100 | 177 | `source:QuillCodeCore` | `Sources/QuillCodeCore/TestCommandLexiconParser.swift` | no automated issues |
+| A+ | 100 | 71 | `source:QuillCodeCore` | `Sources/QuillCodeCore/TestCommandLexiconScope.swift` | no automated issues |
+| A+ | 100 | 91 | `source:QuillCodeCore` | `Sources/QuillCodeCore/TestCommandLexiconShellSegments.swift` | no automated issues |
+| A+ | 100 | 109 | `source:QuillCodeCore` | `Sources/QuillCodeCore/TestCommandLexiconTables.swift` | no automated issues |
 | A+ | 100 | 36 | `source:QuillCodeCore` | `Sources/QuillCodeCore/ThreadEventModels.swift` | no automated issues |
 | A+ | 100 | 129 | `source:QuillCodeCore` | `Sources/QuillCodeCore/ThreadTriageRecords.swift` | no automated issues |
 | A+ | 100 | 65 | `source:QuillCodeCore` | `Sources/QuillCodeCore/ToolCall.swift` | no automated issues |
